More cohesive placing
